requirements.txt
pandas
pandas-gbq
google-cloud-bigquery
airtable-python-wrapper
python-dotenv
pyarrow